% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_A.R
\name{sim_A}
\alias{sim_A}
\title{sim_A - Simulate unweighted network}
\usage{
sim_A(
  N,
  mus,
  omegas,
  p_k,
  beta0,
  model = "NDH",
  precision_R_effects,
  remove_isolates = F
)
}
\arguments{
\item{N}{Number of nodes/actors}

\item{mus}{Cluster centers (KxD matrix)}

\item{omegas}{Precision matrices for cluster (DxDxK array)}

\item{p_k}{Mixture probabilities (Kx1 vector)}

\item{beta0}{Intercept parameter}

\item{model}{Model to simulate data from ("NDH", "RS", or "RSR)}

\item{precision_R_effects}{Precision for random effects (a scalar for "RS" or 2x2 matrix for "RSR")}

\item{remove_isolates}{A logical indicating whether or not to exclude isolates from network}
}
\value{
Add details
}
\description{
sim_A - Simulate unweighted network
}
\details{
More summary needed here.
}
