% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_priors.R
\name{specify_priors}
\alias{specify_priors}
\title{Specify prior hyperparameters for EM algorithm}
\usage{
specify_priors(D, K, model, n_interior_knots = NULL, a, b, c, G, nu, e, f)
}
\arguments{
\item{D}{An integer specifying the dimension of the latent positions.}

\item{K}{An integer specifying the total number of clusters.}

\item{model}{A character string specifying the model:
\itemize{
\item{'NDH': \strong{undirected} network with no degree heterogeneity}
\item{'RS': \strong{undirected} network with degree heterogeneity}
\item{'RSR': \strong{directed} network with degree heterogeneity}
}}

\item{n_interior_knots}{An integer specifying the number of interior knots used in fitting a natural cubic spline for degree heterogeneity models (i.e., 'RS' and 'RSR' only; default is \code{NULL}).}

\item{a}{A numeric vector of length D...}

\item{b}{A numeric value...}

\item{c}{A numeric value...}

\item{G}{A numeric \eqn{N \times D} matrix...}

\item{nu}{A numeric vector of length K...}

\item{e}{A numeric vector. Specifically a vector of length \code{1 + (model =="RS")*(n_interior_knots + 1) +  (model =="RSR")*2*(n_interior_knots + 1)}...}

\item{f}{A numeric square matrix. Specifically dimension \code{1 + (model =="RS")*(n_interior_knots + 1) +  (model =="RSR")*2*(n_interior_knots + 1)}...}
}
\value{
A list of prior hyperparameters for the EM algorithm generated from the input values in a structure accepted by \code{JANE}.
}
\description{
A function that allows the user to specify the prior hyperparameters for the EM algorithm in a structure accepted by \code{JANE}.
}
\details{
For the current implementation we require that all elements of the nu vector be >= 1 to prevent against negative mixture probabilities for empty clusters.
}
